Reorganize navigators
