+++ navigation/FoundDogStack.js
@@ -1,6 +1,5 @@
 import React from 'react'
 import { createStackNavigator } from '@react-navigation/stack'
-import InitAppScreen from '../screens/InitAppScreen'
 import FoundDogScreen from '../screens/FoundDogScreen'
 import PictureScreen from '../screens/camera/PictureScreen'
 import AcceptPictureScreen from '../screens/camera/AcceptPictureScreen'
@@ -9,8 +8,10 @@ import ShowPictureScreen from '../screens/camera/ShowPictureScreen'
 const Stack = createStackNavigator()
 const InitAppStack = () => {
   return (
-    <Stack.Navigator screenOptions={{ headerShown: false }}>
-      <Stack.Screen name="InitApp" component={InitAppScreen} />
+    <Stack.Navigator
+      screenOptions={{ headerShown: false }}
+      initialRouteName="Picture"
+    >
       <Stack.Screen name="Picture" component={PictureScreen} />
       <Stack.Screen name="AcceptPicture" component={AcceptPictureScreen} />
       <Stack.Screen
