+++ screens/ShowFoundDogInfoScreen.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react'
+import React, { useRef } from 'react'
 import { StyleSheet, ScrollView, Dimensions } from 'react-native'
 import PropTypes from 'prop-types'
 import { Icon, ListItem } from 'react-native-elements'
@@ -7,73 +7,68 @@ import Image from 'react-native-scalable-image'
 
 const width = Dimensions.get('window').width
 const height = Dimensions.get('window').height
-export default class ShowFoundDogInfoScreen extends Component {
-  constructor(props) {
-    super(props)
-  }
-  render() {
-    const { dogInfo } = this.props.route.params
-    const parsedDate = new Date(dogInfo.date)
-    const info = [
-      {
-        title: 'Fecha aviso: ' + parsedDate.toDateString(),
-        icon: <Icon name="calendar" type="font-awesome" color="#517fa4" />,
-      },
-      {
-        title: dogInfo.sex,
-        icon: (
-          <Icon
-            name="gender-male-female"
-            type="material-community"
-            color="#517fa4"
-          />
-        ),
-      },
-      {
-        title: dogInfo.commentary,
-        icon: <Icon name="info" type="material" color="#517fa4" />,
-      },
-    ]
-    return (
-      <ScrollView>
-        <Image
-          source={{ uri: dogInfo.imageLinks }}
-          style={styles.image}
-          width={width * 0.6}
+
+const ShowFoundDogInfoScreen = ({ route }) => {
+  const map = useRef(null)
+  const { dogInfo } = route.params
+  const parsedDate = new Date(dogInfo.date)
+  const info = [
+    {
+      title: 'Fecha aviso: ' + parsedDate.toDateString(),
+      icon: <Icon name="calendar" type="font-awesome" color="#517fa4" />,
+    },
+    {
+      title: dogInfo.sex,
+      icon: (
+        <Icon
+          name="gender-male-female"
+          type="material-community"
+          color="#517fa4"
         />
-        {info.map((item, i) => (
-          <ListItem
-            key={i}
-            title={item.title}
-            leftIcon={item.icon}
-            bottomDivider
-          />
-        ))}
-        <MapView
-          style={styles.mapStyle}
-          initialRegion={{
+      ),
+    },
+    {
+      title: dogInfo.commentary,
+      icon: <Icon name="info" type="material" color="#517fa4" />,
+    },
+  ]
+  return (
+    <ScrollView>
+      <Image
+        source={{ uri: dogInfo.imageLinks }}
+        style={styles.image}
+        width={width * 0.6}
+      />
+      {info.map((item, i) => (
+        <ListItem
+          key={i}
+          title={item.title}
+          leftIcon={item.icon}
+          bottomDivider
+        />
+      ))}
+      <MapView
+        style={styles.mapStyle}
+        initialRegion={{
+          latitude: dogInfo.marker.latitude,
+          longitude: dogInfo.marker.longitude,
+          latitudeDelta: 0.0025,
+          longitudeDelta: 0.0025,
+        }}
+        showsBuildings={false}
+        loadingEnabled={true}
+        ref={map}
+      >
+        <Marker
+          coordinate={{
             latitude: dogInfo.marker.latitude,
             longitude: dogInfo.marker.longitude,
-            latitudeDelta: 0.0025,
-            longitudeDelta: 0.0025,
           }}
-          showsBuildings={false}
-          loadingEnabled={true}
-          ref={map => {
-            this.map = map
-          }}
-        >
-          <Marker
-            coordinate={{
-              latitude: dogInfo.marker.latitude,
-              longitude: dogInfo.marker.longitude,
-            }}
-            title={'Acá fue encontrado.'}
-          />
-        </MapView>
-      </ScrollView>
-    )
-  }
+          title={'Acá fue encontrado.'}
+        />
+      </MapView>
+    </ScrollView>
+  )
 }
 
 ShowFoundDogInfoScreen.propTypes = {
@@ -93,3 +88,5 @@ const styles = StyleSheet.create({
     width: width * 0.85,
   },
 })
+
+export default ShowFoundDogInfoScreen
