+++ screens/ShowLostDogInfoScreen.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react'
+import React, { useRef } from 'react'
 import { StyleSheet, ScrollView, Dimensions } from 'react-native'
 import PropTypes from 'prop-types'
 import { Icon, ListItem } from 'react-native-elements'
@@ -7,80 +7,68 @@ import Image from 'react-native-scalable-image'
 
 const width = Dimensions.get('window').width
 const height = Dimensions.get('window').height
-export default class ShowLostDogInfoScreen extends Component {
-  constructor(props) {
-    super(props)
-    this.dogInfo = this.props.route.params.dogInfo
-    this.props.navigation.setOptions({ title: this.dogInfo.name })
-  }
-  // static navigationOptions = ({ navigation }) => {
-  //   return {
-  //     title: navigation.getParam('dogInfo', 'NO-dog').name,
-  //   }
-  // }
-  render() {
-    const dogInfo = this.dogInfo
-    const parsedDate = new Date(dogInfo.date)
-    const info = [
-      {
-        title: 'Fecha aviso: ' + parsedDate.toDateString(),
-        icon: <Icon name="calendar" type="font-awesome" color="#517fa4" />,
-      },
-      {
-        title: dogInfo.sex,
-        icon: (
-          <Icon
-            name="gender-male-female"
-            type="material-community"
-            color="#517fa4"
-          />
-        ),
-      },
-      {
-        title: dogInfo.commentary,
-        icon: <Icon name="info" type="material" color="#517fa4" />,
-      },
-    ]
-    return (
-      <ScrollView>
-        <Image
-          source={{ uri: dogInfo.imageLinks }}
-          style={styles.image}
-          width={width * 0.6}
+const ShowLostDogInfoScreen = ({ navigation, route }) => {
+  const map = useRef(null)
+  const { dogInfo } = route.params
+  navigation.setOptions({ title: dogInfo.name || '' })
+  const parsedDate = new Date(dogInfo.date)
+  const info = [
+    {
+      title: 'Fecha aviso: ' + parsedDate.toDateString(),
+      icon: <Icon name="calendar" type="font-awesome" color="#517fa4" />,
+    },
+    {
+      title: dogInfo.sex,
+      icon: (
+        <Icon
+          name="gender-male-female"
+          type="material-community"
+          color="#517fa4"
         />
-        {info.map((item, i) => (
-          <ListItem
-            key={i}
-            title={item.title}
-            leftIcon={item.icon}
-            bottomDivider
-          />
-        ))}
-        <MapView
-          style={styles.mapStyle}
-          initialRegion={{
+      ),
+    },
+    {
+      title: dogInfo.commentary,
+      icon: <Icon name="info" type="material" color="#517fa4" />,
+    },
+  ]
+  return (
+    <ScrollView>
+      <Image
+        source={{ uri: dogInfo.imageLinks }}
+        style={styles.image}
+        width={width * 0.6}
+      />
+      {info.map((item, i) => (
+        <ListItem
+          key={i}
+          title={item.title}
+          leftIcon={item.icon}
+          bottomDivider
+        />
+      ))}
+      <MapView
+        style={styles.mapStyle}
+        initialRegion={{
+          latitude: dogInfo.marker.latitude,
+          longitude: dogInfo.marker.longitude,
+          latitudeDelta: 0.0025,
+          longitudeDelta: 0.0025,
+        }}
+        showsBuildings={false}
+        loadingEnabled={true}
+        ref={map}
+      >
+        <Marker
+          coordinate={{
             latitude: dogInfo.marker.latitude,
             longitude: dogInfo.marker.longitude,
-            latitudeDelta: 0.0025,
-            longitudeDelta: 0.0025,
           }}
-          showsBuildings={false}
-          loadingEnabled={true}
-          ref={map => {
-            this.map = map
-          }}
-        >
-          <Marker
-            coordinate={{
-              latitude: dogInfo.marker.latitude,
-              longitude: dogInfo.marker.longitude,
-            }}
-            title={'Acá se perdió.'}
-          />
-        </MapView>
-      </ScrollView>
-    )
-  }
+          title={'Acá se perdió.'}
+        />
+      </MapView>
+    </ScrollView>
+  )
 }
 
 ShowLostDogInfoScreen.propTypes = {
@@ -100,3 +88,5 @@ const styles = StyleSheet.create({
     width: width * 0.85,
   },
 })
+
+export default ShowLostDogInfoScreen
